Overwriting main
Merging was causing issues, doing it manually

+++ funcs.py
@@ -1,80 +1,36 @@
-import os,re,time
-from simple_term_menu import TerminalMenu
+import argparse,os,re,time
 from termcolor import colored
+from tools import APT_PACKAGES,GITHUBS,PYPI_PACKAGES
 
+#------- God i'm so sorry -------#
+# This grabs the IP address of tun0 and uses it to start generating malicious binaries
+## TODO: Create a method to select what interface you want to use
+# ip_addr = os.popen('ip addr show tun0 | grep "\\<inet\\>" | awk \'{ print $2 }\' | awk -F "/" \'{ print $1 }\'').read().strip()
+# ip_addr = os.popen('ip addr show eth0 | grep "\\<inet\\>" | awk \'{ print $2 }\' | awk -F "/" \'{ print $1 }\'').read().strip()
+# This port is used for malicious binary generation
+# listen_port = 6969
+#
+#### !!! Consider creating an empty array that we append the finished software to once it has been installed
+#TODO: Grab the neo4j database info, make sure its running and provide to user
+#TODO: Grab the neo4j webserver & the associated port ---> netstat -tano | grep -i "7474"
+#TODO: Adjust the dir/file checks to local, rather than abspath
+#TODO: After install, rename the privilege-escalation-suite dir to PEAS
+#TODO: Ensure that the Metasploit database service is up & running, provide info to the user
+#TODO: Maybe do a search to check if any errors or packages werent able to be added during the script
+#TODO: Potentially rename the functions so they make more sense, and condense
+#TODO: Scrub /etc/hosts file so that it only has the typical localhost/kali entries
+#TODO: Add command-line argument options
+#TODO: Check to make sure the kali installation is a proper VMDK, because the others kinda break
+#TODO: x11 keyboard injection script
+#TODO: Include ansible malicious playbook
+#TODO: Include windows persistence snippet(s)
 
-APT_PACKAGES = [
-	'apt-transport-https',
-	'bloodhound',
-	'chromium',
-	'crackmapexec',
-	'enum4linux',
-	'gobuster',
-	'golang-go',
-	'jxplorer',
-	'metasploit-framework',
-	'nginx',
-	'remmina',
-	'seclists',
-	'smbmap',
-	'snmpcheck',
-	'sshoot',
-	'sshuttle',
-	'subfinder',
-	'sublime-text',
-	'tilix',
-	'wfuzz',
-	'xfreerdp',
-	'yersinia'
-]
+# sudo systemctl start postgresql
+# <check to ensure the service is running now>
+# msfdb init
+# ------------------------------#
 
-#TODO: Alphabetize
-GITHUBS = [
-   'https://github.com/0v3rride/Enum4LinuxPy.git',
-	'https://github.com/21y4d/nmapAutomator.git',
-	'https://github.com/galkan/crowbar.git',
-	'https://github.com/BishopFox/rmiscout.git',
-	'https://github.com/cnotin/SplunkWhisperer2.git',
-	'https://github.com/unode/firefox_decrypt.git',
-	'https://github.com/iagox86/dnscat2.git',
-	'https://github.com/lukebaggett/dnscat2-powershell.git',
-	'https://github.com/darkoperator/Metasploit-Plugins.git',
-	'https://github.com/frohoff/ysoserial.git',
-	'https://github.com/GhostPack/Seatbelt.git',
-	'https://github.com/HackPlayers/evil-winrm.git',
-	'https://github.com/n0b0dyCN/redis-rogue-server.git',
-	'https://github.com/nccgroup/vlan-hopping.git',
-	'https://github.com/danielbohannon/Invoke-Obfuscation.git',
-	'https://github.com/NickstaDB/BaRMIe.git',
-	'https://github.com/samratashok/nishang.git',
-	'https://github.com/p3nt4/Invoke-SocksProxy.git',
-	'https://github.com/silent6trinity/pwnkit.git',
-	'https://github.com/rebootuser/LinEnum.git',
-	'https://github.com/RUB-NDS/PRET.git',
-	'https://github.com/SecureAuthCorp/Impacket.git',
-	'https://github.com/sosdave/KeyTabExtract.git',
-	'https://github.com/vulnersCom/nmap-vulners.git',
-	'https://github.com/int0x33/nc.exe',
-	'https://github.com/danielbohannon/Invoke-Obfuscation.git',
-	'https://github.com/jeroennijhof/vncpwd.git'
-]
-
-#Kerbrute releases
-# https://github.com/ropnop/kerbrute/releases/download/v1.0.3/kerbrute_linux_amd64
-# https://github.com/ropnop/kerbrute/releases/download/v1.0.3/kerbrute_windows_amd64.exe
-
-#pspy release
-#https://github.com/DominicBreuker/pspy/releases/download/v1.2.0/pspy64
-
-PYPI_PACKAGES = [
-	'one-lin3r',
-	'pypykatz',
-	'pygtk',
-	'ptftpd',
-	'bloodhound',
-	'colorama',
-	'pysnmp'
-]
+kit_location = os.getcwd()
 
 # ---- Begin Function declarations -----
 
@@ -83,7 +39,8 @@ def nginx_config():
 	os.system("sudo mkdir -p /var/www/uploads/Exfil")
 	os.system("sudo chown -R www-data:www-data /var/www/uploads/Exfil")
 	os.system("sudo cp ./upload.conf /etc/nginx/sites-available/upload.conf")
-	os.system("sudo ln -s /etc/nginx/sites-available/upload.conf /etc/nginx/sites-enabled/")
+	if not os.path.exists("/etc/nginx/sites-enabled/upload.conf"):
+		os.system("sudo ln -s /etc/nginx/sites-available/upload.conf /etc/nginx/sites-enabled/")
 	os.system("sudo systemctl restart nginx.service")
 	os.system("sudo rm /etc/nginx/sites-enabled/default")
 	# Usage
@@ -114,14 +71,13 @@ def go_install():
 
 #Consider moving into environment setup
 def msfdb_init():
-	#TODO: Check and make sure the msfdb is actually up and running
+	#TODO: Check and make sure the msfdb is actually up and running (msfdb run)
 	os.system('sudo systemctl start postgresql')
 	os.system('systemctl status postgresql')
 	os.system('sudo msfdb init')
 	print("MSF Database Initialized")
 	print("Creating msfconsole.rc file")
-	os.system(f' cp ./msfconsole.rc /home/kali/.msf4/msfconsole.rc')
-
+	os.system(f' cp ./msfconsole.rc /home/kali/.msf4/msfconsole.rc') #This currently doesn't work due to the dirs switching around
 
 #Consider moving into environment setup
 def neo4j_init():
@@ -137,9 +93,9 @@ def neo4j_init():
 # This whole PEAS mess needs to be fixed later
 def peas_download():
 	# For the time being - just scrub the PEAS directory and re-obtain
-	if os.path.exists("/opt/PEAS"):
+	if os.path.exists("./PEAS"):
 		#Lol, risky
-		os.system("sudo rm -rf /opt/PEAS")
+		os.system("sudo rm -rf ./PEAS")
 		grab_peas()
 	else:
 		grab_peas()
@@ -148,11 +104,10 @@ def grab_peas():
 	linpeas_sh = 'https://github.com/carlospolop/PEASS-ng/releases/download/20221009/linpeas.sh'
 	winpeas_bat = 'https://github.com/carlospolop/PEASS-ng/releases/download/20221009/winPEAS.bat'
 	winpeas_exe = 'https://github.com/carlospolop/PEASS-ng/releases/download/20221009/winPEASany.exe'
-	os.system(f"sudo mkdir /opt/PEAS")
-	os.system(f"sudo wget {linpeas_sh} -qO /opt/PEAS/linpeas.sh ; sudo chmod +x /opt/PEAS/linpeas.sh")
-	os.system(f"sudo wget {winpeas_bat} -qO /opt/PEAS/winpeas.bat")
-	os.system(f"sudo wget {winpeas_exe} -qO /opt/PEAS/winpeas.exe")
-
+	os.system(f"sudo mkdir ./PEAS")
+	os.system(f"sudo wget {linpeas_sh} -qO ./PEAS/linpeas.sh ; sudo chmod +x ./PEAS/linpeas.sh")
+	os.system(f"sudo wget {winpeas_bat} -qO ./PEAS/winpeas.bat")
+	os.system(f"sudo wget {winpeas_exe} -qO ./PEAS/winpeas.exe")
 
 
 def shell_creation():
@@ -171,46 +126,19 @@ def shell_creation():
 # os.system("ln -s /opt/nmapAutomator/nmapAutomator.sh /usr/local/bin/ && chmod +x /opt/nmapAutomator/nmapAutomator.sh")
 # sudo ln -s /opt/LinEnum.sh /usr/local/bin/'
 # sudo ln -s /opt/.local/bin/one-lin3r /usr/local/bin
-def tool_install():
-	#Temp method to grab lazagne and the old firefox decrypt for python2
-	lazagne_exe = 'https://github.com/AlessandroZ/LaZagne/releases/download/2.4.3/lazagne.exe'
-	os.system(f"sudo wget {lazagne_exe} -qO /opt/lazagne.exe")
-	ff_decrypt_old = 'https://github.com/unode/firefox_decrypt/archive/refs/tags/0.7.0.zip'
-	os.system(f"sudo wget {ff_decrypt_old} -qO /opt/FirefoxDecrypt_ForPython2")
-	
-	#### END TEMP METHOD
-	
-	def is_repo_installed(repo_url):
-		if a_match := re.match(r"https://.+/(.+)\.git", repo_url):
-			return os.path.exists(f"/opt/{a_match.group(1)}")
+
+def structure_setup():
+	"""Meant to create directory structure and organize tools into"""
+	#NOTE-> THIS IS TESTED AND WORKS PROPERLY
+	DIRS = ["Linux","Windows","ActiveDirectory","C2Frameworks"]
+	for dir in DIRS:
+		if os.path.exists(f"/home/kali/Downloads/{dir}"):
+			print(f"{dir} FOLDER EXISTS")
 		else:
-			print(colored(f'INVALID URL: {repo_url}', 'red'))
-			# Returning True here because if the url isn't valid, then we definitely don't want to try installing
-			return True
+			os.mkdir(f"/home/kali/Downloads/{dir}")
+			print(f"created the {dir} directory")
 
-	for git_url in GITHUBS:
-		print(f"Checking for local install of: {git_url}")
-		if is_repo_installed(git_url):
-			print(colored(f"Found in /opt continuing...\n"))
-		else:
-			os.system(f"git clone {git_url}")
-			print(colored("Repo cloned! Moving on...\n", "green"))
-			#return()
-
-	# begin installing pypi & apt packages
-	for pkg in APT_PACKAGES:
-		os.system(f'sudo apt install {pkg} -y 1>/dev/null')
-		os.system('sudo apt install -y 1>/dev/null')
-		print(colored(f'APT {pkg} successfully installed by script', "green"))
-	for pkg in PYPI_PACKAGES:
-		os.system(f'pip3 install {pkg} 1>/dev/null')
-		print(colored(f'PYPI {pkg} successfully installed by script', "green"))
-	peas_download()
-	os.system("sudo ln -s /opt/nmapAutomator/nmapAutomator.sh /usr/local/bin/ && chmod +x /opt/nmapAutomator/nmapAutomator.sh")
-	print("tool_install() Completed")
-	return True
-
-
+# This could potentially be thrown away now and just use the .deb install now
 def sublime_download():
 	sublime = 'deb https://download.sublimetext.com/ apt/stable/'
 	os.system('sudo wget -qO - https://download.sublimetext.com/sublimehq-pub.gpg\
@@ -234,49 +162,11 @@ def system_update():
 	print(colored("Finished SYSTEM setup", 'green'))
 	return()
 
-def terminal_selection():
-	""" This is what is used within main() to control the function flow """
-	main_menu_title = "Automated Kit Buildout Script, Select ALL, TOOLS, SHELL or TEST\n"
-	main_menu_cursor = "-> "
-
-	options = ["ALL", "TOOLS", "SHELL (BROKEN)", "TEST"]
-	# begin TUI Custom configuration(s)
-	terminal_menu = TerminalMenu(
-		options,
-		title=main_menu_title,
-		menu_cursor=main_menu_cursor)
-	menu_entry_index = terminal_menu.show()
-	user_selection = {options[menu_entry_index]}
-	# Choice menu
-	if menu_entry_index == 0:
-		print(("Match Successful on ALL"))
-		nginx_config()
-		system_update()
-		msfdb_init()
-		neo4j_init()
-		#software_update()
-	elif menu_entry_index == 1:
-		print("Match successful on TOOLS")
-		#software_update()
-		nginx_config()
-		tool_install()
-		tool_update()
-		msfdb_init()
-		neo4j_init()
-		jon()
-		#go_install()
-	elif menu_entry_index == 2:
-		print("Match successful on SHELL")
-		shell_creation()
-	elif menu_entry_index == 3:
-		test()
-	else:
-		print("Match failed.")
-
 def test():
-	peas_download()
 	print(os.getlogin()) # Interestingly enough - this returns the actual user
+	print(f"Kit.py Location: {kit_location}")
 	print(os.system("whoami")) # This returns as root (since it's run as sudo)
+	print("Test Completed")
 	return()
 
 def jon():
@@ -311,4 +201,43 @@ def tool_update():
 	print(colored("Finished locate DB update \n", 'green'))
 	nmap_update()
 	rockyou()
+	return True
+
+def tool_install():
+	os.chdir("/home/kali/Downloads")
+	####Temp method to grab lazagne and the old firefox decrypt for python2
+	lazagne_exe = 'https://github.com/AlessandroZ/LaZagne/releases/download/2.4.3/lazagne.exe'
+	os.system(f"sudo wget {lazagne_exe} -qO ./lazagne.exe")
+	ff_decrypt_old = 'https://github.com/unode/firefox_decrypt/archive/refs/tags/0.7.0.zip'
+	os.system(f"sudo wget {ff_decrypt_old} -qO ./FirefoxDecrypt_ForPython2")
+	#### END TEMP METHOD
+	
+	def is_repo_installed(repo_url):
+		if a_match := re.match(r"https://.+/(.+)\.git", repo_url):
+			return os.path.exists(f"./{a_match.group(1)}")
+		else:
+			print(colored(f'INVALID URL: {repo_url}', 'red'))
+			# Returning True here because if the url isn't valid, then we definitely don't want to try installing
+			return True
+
+	for git_url in GITHUBS:
+		print(f"Checking for local install of: {git_url}")
+		if is_repo_installed(git_url):
+			print(colored(f"Found in current directory, continuing...\n"))
+		else:
+			os.system(f"git clone {git_url}")
+			print(colored("Repo cloned! Moving on...\n", "green"))
+			#return()
+
+	# begin installing pypi & apt packages
+	for pkg in APT_PACKAGES:
+		os.system(f'sudo apt install {pkg} -y 1>/dev/null')
+		os.system('sudo apt install -y 1>/dev/null')
+		print(colored(f'APT {pkg} successfully installed by script', "green"))
+	for pkg in PYPI_PACKAGES:
+		os.system(f'pip3 install {pkg} 1>/dev/null')
+		print(colored(f'PYPI {pkg} successfully installed by script', "green"))
+	peas_download()
+	os.system("sudo ln -s ./nmapAutomator/nmapAutomator.sh /usr/local/bin/ && chmod +x ./nmapAutomator/nmapAutomator.sh")
+	print("tool_install() Completed")
 	return True
